+++ config/flare.php
@@ -28,6 +28,7 @@ return [
         'anonymize_ips' => true,
         'collect_git_information' => true,
         'report_queries' => true,
+        'maximum_number_of_collected_queries' => 200,
         'report_query_bindings' => true,
         'report_view_data' => true,
     ],