+++ tests/stubs/views/undefined-variable-1.blade.php
@@ -1,0 +1,11 @@
+{{-- Intentional typo for test --}}
+
+This is some contents
+
+<footer>{{ $footerDescriptin }}</footer>
+
+@isset($something)
+    {{ $something }}
+@endisset
+
+Test